Added README.md at root and in build
README.md in the root of the repo serves as an introduction to cgcs_vector_ptr. It's still a work in progress. Also, we have a README.md in the build directory. I explain what the subdirectories in the build directory are for -- they are provided for convenience when using CMake. (since I am lazy and would rather not have to make them when CMake tells me a directory does not exist)