using_cgcs_vptr and CGCS_VECTOR_PTR_SHORTNAMES changed to macro function calls, and init/deinit functions no longer return.
The shortened names that are used for cgcs_vptr_t should only be recognized when they are in the form of a function call --
previously, they were not.

Also, the init/deinit functions had previously returned (cgcs_vptr_t *), now they return as (void).
Because of the changes to init/deinit, cgcs_vptr_new and cgcs_vptr_delete are now inline functions.