Restructured repo. Library is in ./src, demo is in ./demo. Using CMake now.
This repository is designed to be a submodule in other repositories that want to use cgcs_vector. The repo now has a top-level CMakeLists.txt, that adds ./src and ./demo as subdirectories.

The ./src folder has the cgcs_vector library itself, which is what other projects will specify when wanting to use cgcs_vector. It has its own CMakeLists.txt.

The ./demo folder has test code for cgcs_vector, to sample how it can be used. It too also has its own CMakeLists.txt.